Add unit tests for UserFollowUseCase.kt

+++ libraries/data-domain/src/main/java/com/raudonikis/data_domain/user_follow/UserFollowUseCase.kt
@@ -1,36 +1,44 @@
 package com.raudonikis.data_domain.user_follow
 
 import com.raudonikis.common.Outcome
+import com.raudonikis.core.providers.di.IODispatcher
 import com.raudonikis.data_domain.user.User
 import com.raudonikis.data_domain.user.cache.UserDao
 import com.raudonikis.network.GamesApi
 import com.raudonikis.network.utils.safeNetworkResponse
+import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 class UserFollowUseCase @Inject constructor(
     private val gamesApi: GamesApi,
-    private val userDao: UserDao
+    private val userDao: UserDao,
+    @IODispatcher
+    private val ioDispatcher: CoroutineDispatcher,
 ) {
 
     suspend fun followUser(user: User): Outcome<List<Nothing>> {
-        return withContext(Dispatchers.IO) {
+        return withContext(ioDispatcher) {
             safeNetworkResponse {
                 gamesApi.followUser(user.uuid)
             }
-        }.toOutcome().onSuccess {
-            userDao.addNewFollowingUser(user)
+        }.toOutcome().also { outcome ->
+            if (outcome is Outcome.Success || outcome is Outcome.Empty) {
+                userDao.addNewFollowingUser(user)
+            }
         }
     }
 
     suspend fun unfollowUser(user: User): Outcome<List<Nothing>> {
-        return withContext(Dispatchers.IO) {
+        return withContext(ioDispatcher) {
             safeNetworkResponse {
                 gamesApi.unfollowUser(user.uuid)
             }
-        }.toOutcome().onSuccess {
-            userDao.removeFollowingUser(user)
+        }.toOutcome().also { outcome ->
+            if (outcome is Outcome.Success || outcome is Outcome.Empty) {
+                userDao.removeFollowingUser(user)
+            }
         }
     }
 }
